+++ src/Reverie/revcb/controllerBuilder.ts
@@ -6,6 +6,7 @@ import { RevCBController } from './RevCBController';
 import { setupLightEntities } from './setupLightEntities';
 import { setupMemoryTracking } from './setupMemoryTracking';
 import { setupMotorEntities } from './setupMotorEntities';
+import { setupPositionInputs } from './setupPositionInputs';
 import { setupPositionSensors } from './setupPositionSensors';
 import { setupPresetButtons } from './setupPresetButtons';
 
@@ -56,6 +57,7 @@ export const controllerBuilder = async (mqtt: IMQTTConnection, deviceData: IDevi
   setupLightEntities(mqtt, controller);
   setupMotorEntities(mqtt, controller);
   setupPositionSensors(mqtt, controller);
+  setupPositionInputs(mqtt, controller);
   setupMemoryTracking(mqtt, controller);
 
   return controller;